page display works

+++ data/src/main/java/com/example/data/index/entities/IndexValue.kt
@@ -3,6 +3,6 @@ package com.example.data.index.entities
 data class IndexValue(
     val researchId: Long,
     val indexNameId: Long,
-    val value: Int,
+    val value: Double,
     val waterQuality: String
 )
